third_party/blink/web_tests/fast/css/hover-active-quirks-expected.txt

Quirks mode - matching :hover and :active

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.eventSender is defined.
PASS testElement(t1.firstChild, true, false) is green
PASS testElement(t2.firstChild, true, true) is green
PASS testElement(t3.firstChild, true, true) is green
PASS testElement(t4.firstChild, true, false) is green
PASS testElement(t5.firstChild, true, true) is green
PASS testElement(t6.firstChild, true, true) is green
PASS testElement(t7.firstChild, false, false) is green
PASS testElement(t8.firstChild, false, false) is green
PASS testElement(t9.firstChild, true, false) is green
PASS testElement(t10.firstChild, true, false) is green
PASS testElement(t11.firstChild, false, true) is green
PASS testElement(t12.firstChild, false, true) is green
PASS testElement(t13, true, false) is green
PASS testElement(t14, true, false) is green
PASS testElement(t15, true, false) is green
PASS testElement(t16, true, false) is green
PASS testElement(t17, true, false) is green
PASS testElement(t18, true, false) is green
PASS testElement(t19.firstChild, true, false) is green
PASS testElement(t20, true, true) is green
PASS testElement(t21, true, true) is green
PASS testElement(t22, true, true) is green
PASS testElement(t23, true, true) is green
PASS testElement(t24, true, true) is green
PASS testElement(t25, true, true) is green
PASS testElement(t26.firstChild, true, true) is green
PASS successfullyParsed is true

TEST COMPLETE


Should stay green when hovered and/or active: 
 
 
 
 
 

Should be green when not hovered: 

Should be green when not active: 

Should be green when hovered and not active: 
 

Should be green when active and not hovered: 
 

Should be green when hovered:       

Should be green when active:       
